add docs to properties datatype list (#150)
This list was automatically generated and then manually reformatted. This process is being documented so it can be repeated consistently and be improved upon.

Currently, this documentation is in the only file of this type. Should this file be moved to its own directory at some point, it might make sense to move this documentation to its own README.md

Bug: T261295