Add ratio-based references features
These are useful in case of finding how many of statements have proper
references and if an item is properly referenced or not.

There's a feature that hold the total number of references but it's not
useful and I'm not sure the model combine it properly with the total
number of claims

_filter_nonexternal_identifier_statements is called multiple times while
evaluating a single entity. It might be worthwhile to look into whether
using an LRUCache here would improve performance.