Split important languages to have their own feature
Currently, if an item has 4 out of 10 important languages as label,
they'll get a feature like "4" which reduces the good predictors like
the difference between English and Chinese. With this change they get a
bigger set of features in a one-hot encoding way. For example
[0, 0, 1, 0, 1, 0, 1, 0, ...]

This increases number of correct predictions on the labelled dataset
from 4175 to 4198 (out of 4951).