Remove impactless property suggester feature
It turned out, that this feature has no significant effect on the
scoring of an item revision. And what slight effect it has seemed to be
minimally negative.

In addition, it comes with significant drawbacks as it requires live
requests to the wikidata API during training and scoring via the API.
Those requests are too numerous while processing dumps, so they were
always skipped there. This meant that revisions could theoretically have
had a different score when scored via the API or based on dumps, which
is also undesirable.

Bug: https://phabricator.wikimedia.org/T261850